Rename sbtireader.py to sbti.py
